+++ src/auth.d.ts
@@ -1,0 +1,11 @@
+import type { Session } from "@auth/core/types";
+
+declare module "@auth/core/types" {
+  interface Session {
+    user: {
+      name?: string | null;
+      email?: string | null;
+      providerAccountId?: string | null;
+    };
+  }
+}
