+++ src/env.d.ts
@@ -3,9 +3,6 @@
 /// <reference path="../.astro/types.d.ts" />
 /// <reference types="astro/client-image" />
 
-// eslint-disable-next-line @typescript-eslint/no-unused-vars
-import type { Session } from "@auth/core/types";
-
 interface ImportMetaEnv {
   // Auth.js
   readonly AUTH_SECRET: string;
@@ -22,14 +19,4 @@ interface ImportMetaEnv {
 
 interface ImportMeta {
   readonly env: ImportMetaEnv;
-}
-
-declare module "@auth/core/types" {
-  interface Session {
-    user: {
-      name?: string | null;
-      email?: string | null;
-      providerAccountId?: string | null;
-    };
-  }
 }
